flex order allows flexibility
Let's say that the web looks great on desktop but when trying to go mobile the html code has to be changed and this is a problem. With flexbox order one can re-arrange the order of the elements by using flex order and not having to change the HTML order.