flex alignment + centering
justify-content: flex-start is the default property and value. justify-content: flex-end => moves items within the flex container to the end of the main axis. justify-content: center; moves items to the center of the main axis. justify-content: space-between; divides the space in between for each element. justify-content: space-around; takes into account the space in the left and right of the main axis.